backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
crewai==0.80.0
python-dotenv==1.0.1
groq>=0.4.0
PyMuPDF>=1.23.0
google-generativeai>=0.4.0
